pyproject.toml
[tool.poetry]
name = "vlm_finetune"
version = "0.1.0"
description = "Fine-tuning vision-language models with CUDA"
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.13"

accelerate = "1.10.1"
aiohappyeyeballs = "2.6.1"
aiohttp = "3.13.1"
aiosignal = "1.4.0"
anyio = "4.11.0"
asttokens = "3.0.0"
attrs = "25.4.0"
bert-score = "0.3.13"
bitsandbytes = "0.48.1"
certifi = "2025.10.5"
charset-normalizer = "3.4.4"
colorama = "0.4.6"
comm = "0.2.3"
contourpy = "1.3.3"
cycler = "0.12.1"
datasets = "4.2.0"
debugpy = "1.8.17"
decorator = "5.2.1"
dill = "0.4.0"
evaluate = "0.4.6"
executing = "2.2.1"
filelock = "3.19.1"
fonttools = "4.60.1"
frozenlist = "1.8.0"
fsspec = "2025.9.0"
h11 = "0.16.0"
httpcore = "1.0.9"
httpx = "0.28.1"
huggingface-hub = "0.35.3"
idna = "3.11"
jedi = "0.19.2"
jinja2 = "3.1.6"
kiwisolver = "1.4.9"
markupsafe = "2.1.5"
matplotlib = "3.10.7"
matplotlib-inline = "0.1.7"
mpmath = "1.3.0"
multidict = "6.7.0"
multiprocess = "0.70.16"
nest-asyncio = "1.6.0"
networkx = "3.5"
numpy = "2.3.3"
packaging = "25.0"
pandas = "2.3.3"
parso = "0.8.5"
peft = "0.17.1"
pillow = "11.3.0"
platformdirs = "4.5.0"
prompt-toolkit = "3.0.52"
propcache = "0.4.1"
psutil = "7.1.0"
pure-eval = "0.2.3"
pyarrow = "21.0.0"
pygments = "2.19.2"
pyparsing = "3.2.5"
python-dateutil = "2.9.0.post0"
pytz = "2025.2"
pyyaml = "6.0.3"
pyzmq = "27.1.0"
regex = "2025.9.18"
requests = "2.32.5"
safetensors = "0.6.2"
sentencepiece = "0.2.1"
six = "1.17.0"
sniffio = "1.3.1"
stack-data = "0.6.3"
sympy = "1.13.1"
tokenizers = "0.22.1"
tornado = "6.5.2"
tqdm = "4.67.1"
traitlets = "5.14.3"
transformers = "4.57.1"
typing-extensions = "4.15.0"
tzdata = "2025.2"
urllib3 = "2.5.0"
wcwidth = "0.2.14"
widgetsnbextension = "4.0.14"
xxhash = "3.6.0"
yarl = "1.22.0"

# PyTorch и torchvision с CUDA 12.1
torch = { version = "2.5.1", source = "pytorch-cu121" }
torchvision = { version = "0.20.1", source = "pytorch-cu121" }

[tool.poetry.group.dev.dependencies]
ipykernel = "*"
ruff = "^0.14.2"

[[tool.poetry.source]]
name = "pytorch-cu121"
url = "https://download.pytorch.org/whl/cu121"
priority = "explicit"

[[tool.poetry.source]]
name = "pypi"
priority = "primary"

[tool.ruff]
include = ["*.py", "*.pyi"]
exclude = ["build", "dist", ".venv"]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "F",   # pyflakes
    "I",   # import sorting (аналог isort)
    "UP",  # pyupgrade
    "B",   # bugbear
    "C4",  # comprehensions
]

ignore = ["E501"]

[tool.ruff.lint.isort]
known-first-party = ["vlm_finetune"]
combine-as-imports = true
